requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
et_xmlfile==2.0.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
